[AIRFLOW-5351] Move all GCP Cloud SQL tests in 1 file
